refactor worktree cleanup for orphaned and externally deleted worktrees
