+++ crates/core/src/execution_monitor.rs
@@ -310,233 +310,8 @@ pub async fn cleanup_task_worktrees(
     Ok(())
 }
 
-/// Defensively check for externally deleted worktrees and mark them as deleted in the database
-async fn check_externally_deleted_worktrees(pool: &sqlx::SqlitePool) {
-    let active_attempts = match TaskAttempt::find_by_worktree_deleted(pool).await {
-        Ok(attempts) => attempts,
-        Err(e) => {
-            tracing::error!(
-                "Failed to query active task attempts for external deletion check: {}",
-                e
-            );
-            return;
-        }
-    };
-    tracing::debug!(
-        "Checking {} active worktrees for external deletion...",
-        active_attempts.len()
-    );
-    for (attempt_id, worktree_path) in active_attempts {
-        // Check if worktree directory exists
-        if !std::path::Path::new(&worktree_path).exists() {
-            // Worktree was deleted externally, mark as deleted in database
-            if let Err(e) =
-                db::models::task_attempt::TaskAttempt::mark_worktree_deleted(pool, attempt_id).await
-            {
-                tracing::error!(
-                    "Failed to mark externally deleted worktree as deleted for attempt {}: {}",
-                    attempt_id,
-                    e
-                );
-            } else {
-                tracing::info!(
-                    "Marked externally deleted worktree as deleted for attempt {} (path: {})",
-                    attempt_id,
-                    worktree_path
-                );
-            }
-        }
-    }
-}
-
-/// Find and delete orphaned worktrees that don't correspond to any task attempts
-async fn cleanup_orphaned_worktrees(db: &DBService) {
-    // Check if orphan cleanup is disabled via environment variable
-    if std::env::var("DISABLE_WORKTREE_ORPHAN_CLEANUP").is_ok() {
-        tracing::debug!("Orphan worktree cleanup is disabled via DISABLE_WORKTREE_ORPHAN_CLEANUP environment variable");
-        return;
-    }
-    // todo
-    let worktree_base_dir = PathBuf::from("/path/to/worktree/base");
-
-    // Check if base directory exists
-    if !worktree_base_dir.exists() {
-        tracing::debug!(
-            "Worktree base directory {} does not exist, skipping orphan cleanup",
-            worktree_base_dir.display()
-        );
-        return;
-    }
-
-    // Read all directories in the base directory
-    let entries = match std::fs::read_dir(&worktree_base_dir) {
-        Ok(entries) => entries,
-        Err(e) => {
-            tracing::error!(
-                "Failed to read worktree base directory {}: {}",
-                worktree_base_dir.display(),
-                e
-            );
-            return;
-        }
-    };
-
-    let mut orphaned_count = 0;
-    let mut checked_count = 0;
-
-    for entry in entries {
-        let entry = match entry {
-            Ok(entry) => entry,
-            Err(e) => {
-                tracing::warn!("Failed to read directory entry: {}", e);
-                continue;
-            }
-        };
-
-        let path = entry.path();
-
-        // Only process directories
-        if !path.is_dir() {
-            continue;
-        }
-
-        let worktree_path_str = path.to_string_lossy().to_string();
-        checked_count += 1;
-
-        if let Ok(false) = TaskAttempt::container_ref_exists(&db.pool, &worktree_path_str).await {
-            // This is an orphaned worktree - delete it
-            tracing::info!("Found orphaned worktree: {}", worktree_path_str);
-
-            // For orphaned worktrees, we try to clean up git metadata if possible
-            // then remove the directory
-            if let Err(e) = cleanup_orphaned_worktree_directory(&path).await {
-                tracing::error!(
-                    "Failed to remove orphaned worktree {}: {}",
-                    worktree_path_str,
-                    e
-                );
-            } else {
-                orphaned_count += 1;
-                tracing::info!(
-                    "Successfully removed orphaned worktree: {}",
-                    worktree_path_str
-                );
-            }
-        }
-    }
-
-    if orphaned_count > 0 {
-        tracing::info!(
-            "Cleaned up {} orphaned worktrees (checked {} total directories)",
-            orphaned_count,
-            checked_count
-        );
-    } else {
-        tracing::debug!(
-            "No orphaned worktrees found (checked {} directories)",
-            checked_count
-        );
-    }
-}
-
-/// Clean up an orphaned worktree directory, attempting to clean git metadata if possible
-async fn cleanup_orphaned_worktree_directory(
-    worktree_path: &std::path::Path,
-) -> Result<(), std::io::Error> {
-    // Use WorktreeManager for proper cleanup - it will try to infer the repo path
-    // and clean up what it can, even if the main repo is gone
-    match WorktreeManager::cleanup_worktree(worktree_path, None).await {
-        Ok(()) => {
-            tracing::debug!(
-                "WorktreeManager successfully cleaned up orphaned worktree: {}",
-                worktree_path.display()
-            );
-        }
-        Err(e) => {
-            tracing::warn!(
-                "WorktreeManager cleanup failed for orphaned worktree {}: {}",
-                worktree_path.display(),
-                e
-            );
-
-            // If WorktreeManager cleanup failed, fall back to simple directory removal
-            // This ensures we delete as much as we can
-            if worktree_path.exists() {
-                tracing::debug!(
-                    "Falling back to simple directory removal for orphaned worktree: {}",
-                    worktree_path.display()
-                );
-                std::fs::remove_dir_all(worktree_path).map_err(|e| {
-                    std::io::Error::new(
-                        e.kind(),
-                        format!("Failed to remove orphaned worktree directory: {}", e),
-                    )
-                })?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
 pub async fn execution_monitor(deployment: DeploymentImpl) {
     // let mut interval = tokio::time::interval(tokio::time::Duration::from_secs(5));
-    let mut cleanup_interval = tokio::time::interval(tokio::time::Duration::from_secs(1800)); // 30 minutes
-
-    loop {
-        cleanup_interval.tick().await;
-        tracing::info!("Starting periodic worktree cleanup...");
-
-        // First, defensively check for externally deleted worktrees
-        check_externally_deleted_worktrees(&deployment.db().pool).await;
-
-        // Then, find and delete orphaned worktrees that don't belong to any task
-        cleanup_orphaned_worktrees(&deployment.db()).await;
-
-        // Then, proceed with normal expired worktree cleanup
-        match TaskAttempt::find_expired_for_cleanup(&deployment.db().pool).await {
-            Ok(expired_attempts) => {
-                if expired_attempts.is_empty() {
-                    tracing::debug!("No expired worktrees found");
-                } else {
-                    tracing::info!(
-                        "Found {} expired worktrees to clean up",
-                        expired_attempts.len()
-                    );
-                    for (attempt_id, worktree_path, git_repo_path) in expired_attempts {
-                        if let Err(e) =
-                            delete_worktree(worktree_path, git_repo_path, attempt_id).await
-                        {
-                            tracing::error!(
-                                "Failed to cleanup expired worktree {}: {}",
-                                attempt_id,
-                                e
-                            );
-                        } else {
-                            // Mark worktree as deleted in database after successful cleanup
-                            if let Err(e) =
-                                db::models::task_attempt::TaskAttempt::mark_worktree_deleted(
-                                    &deployment.db().pool,
-                                    attempt_id,
-                                )
-                                .await
-                            {
-                                tracing::error!("Failed to mark worktree as deleted in database for attempt {}: {}", attempt_id, e);
-                            } else {
-                                tracing::info!(
-                                    "Successfully marked worktree as deleted for attempt {}",
-                                    attempt_id
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                tracing::error!("Failed to query expired task attempts: {}", e);
-            }
-        }
-    }
 }
 
 // /// Handle setup script completion
